Let's goooo

+++ jeeapp/client/src/main/java/entities/Currency.java
@@ -3,9 +3,9 @@ package entities;
 public class Currency {
 
     private String name;
-    private Long toEuro;
+    private Double toEuro;
 
-    public Currency(String name, Long toEuro) {
+    public Currency(String name, Double toEuro) {
         this.name = name;
         this.toEuro = toEuro;
     }
@@ -18,11 +18,11 @@ public class Currency {
         this.name = name;
     }
 
-    public Long getToEuro() {
+    public Double getToEuro() {
         return toEuro;
     }
 
-    public void setToEuro(Long toEuro) {
+    public void setToEuro(Double toEuro) {
         this.toEuro = toEuro;
     }
